news_crew/pyproject.toml
[project]
name = "crewai"
version = "0.1.0"
description = "A CrewAI project"
requires-python = ">=3.8"
dependencies = [
    "crewai",
    "python-dotenv",
    "openai",
    "exa-py",
    "streamlit",
    "pyyaml",
    "datetime",
    "string",
    "time"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
